[refac] notification 로직 분리 및 파일로 분류
